requirements.txt
# Core Dependencies
crewai==1.5.0
python-dotenv==1.0.0
google-generativeai==0.4.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.0.1

# Machine Learning & Search
sentence-transformers==3.0.0
faiss-cpu==1.13.0
numpy==1.24.3
torch==2.0.1
chromadb==0.4.24

# Utilities
requests==2.31.0
